Remove old Controllers
