+++ Interfaces/ISettingsService.cs
@@ -1,4 +1,4 @@
-﻿using NanoTwitchLeafs.Controller;
+﻿
 using NanoTwitchLeafs.Objects;
 
 namespace NanoTwitchLeafs.Interfaces;
